build: clear the clang warning backlog
A clang Release build emitted 201 warning lines. Two of them were the warning set asking for
something a Qt/C++20 project cannot give; the rest were real.

Flags (cmake/compiler_warnings.cmake):

  - Drop -Wc++20-compat-pedantic. It warns about constructs incompatible with standards *before*
    C++20, and every one of its ~133 hits was Qt's Q_CONSTINIT expanding to constinit in generated
    moc translation units. Nothing here can act on it.
  - Add -Wno-keyword-macro. Its 7 hits are all the deliberate `#define private public` the test
    suites use to reach privates. The warning rides on -Wpedantic, so it cannot be narrowed to
    non-test code without per-site pragmas.

TimeStamp.h adds -Wdouble-promotion to the pragma block that already wraps the boost/date_time
include: lexical_cast's `const double val_as_double = val;` is instantiated from our translation
unit, so -isystem does not reach it.

The 50 -Wdouble-promotion hits in the graph code came from float Vec2f/QVector2D/QVector4D feeding
Qt's qreal geometry API. Rather than add 50 static_cast<qreal> to code already dense with
static_cast<float>, the arithmetic now happens in qreal from the start -- QtGraphNodeQualifier is a
net 9 lines shorter because converting the position once via toPointF() deleted all 14 of its
existing float casts. Integer truncation that was load-bearing (arrowOffset, height / 2) stays int.

SourceLocationFile was the last caller of the two [[deprecated]] SourceLocation constructors, so
both are gone. Routing it through the factories needed a fix first: the deprecated two-argument
constructor set the start->end back-link via pOther->setOtherLocation(this), but
createEndSourceLocation only set end->start. It now sets both, and SourceLocationTestSuite asserts
the direction that was previously untested.

The rest are one-liners: a loop that ran at most once, an unused lambda capture, a qreal used as a
0-sentinel where the paired pointer says the same thing, an unread constant, and a test argv that
only needed to be const char*.

Also folded in the in-flight warning work already in the tree: the LLVM PointerUnion is/get to
isa/cast migration, shadowed constructor parameters, and assorted sign-conversion fixes.

Verified with the clang_release_build_cxx preset: 0 warnings, 1221/1221 tests pass.

+++ src/features/graph/ui/graph/QtGraphNodeAccess.cpp
@@ -78,7 +78,7 @@ void QtGraphNodeAccess::updateStyle() {
     m_text->setPos(
         static_cast<qreal>(style.textOffset.x + static_cast<float>(m_accessIconSize) + 3.0f),
         static_cast<qreal>(style.textOffset.y + static_cast<float>(m_accessIconSize) - static_cast<float>(style.fontSize)));
-    m_accessIcon->setPos(style.textOffset.x, style.textOffset.y);
+    m_accessIcon->setPos(static_cast<qreal>(style.textOffset.x), static_cast<qreal>(style.textOffset.y));
 
     m_accessIcon->setPixmap(utility::colorizePixmap(m_accessIcon->pixmap(), style.color.icon.c_str()));
   } else {
